requirements_original.txt
flask==3.1.1
flask-cors==6.0.1
flask-sqlalchemy==3.1.1
openai==1.95.1
requests==2.32.3
sqlalchemy==2.0.41
pillow==10.4.0
werkzeug==3.1.3
beautifulsoup4==4.12.3
pyjwt==2.8.0
pydantic==2.5.0

